Build: 1차 프로젝트 배포  (#50)
* Feat : 공통 스타일 생성

* Feat: router 추가

router 추가 및 기본 page 추가

* Feat: Date 컴포넌트, RecipientName 컴포넌트 디자인 시스템 구성

* Feat: 뱃지, 토스트 atoms추가

* Feat: Avatar atoms 추가

Avatar 및 DefaultAvatar 추가

* Feat: TextFeildInput 구현

* Feat: Dropdown 구현

* Remove: test페이지 삭제

* Fix: Dropdown 기능 재구현

* Style: 디렉토리 구조 및 네이밍 변경

* Feat: 뱃지,이모지 컴포넌트 디자인 추가 구현 (#4)

* Feat: GNB  컴포넌트 디자인 및 기능 구현 (#5)

* Feat: GNB 컴포넌트 디자인 및 기능 구현

* Style: GNB 컴포넌트 관련 파일 디렉토리 주소 수정

* Remove: postIdPage.jsx 삭제

* Feat: Button 컴포넌트 기본 타입 구현 (#6)

* Style: eslint rules "off" 옵션 추가

* Feat: Button 컴포넌트  기본 타입 구현

* Feat: Option 컨포넌트 디자인 및 기능 구현 (#7)

* Feat: ContributorStats 컴포넌트 디자인 및 기능 구현 (#8)

* Feat: api 추가

* Feat: Avatars 추가

Senders, Card 등에 사용되는 Avatars 추가

* Feat: ContributorStats 컴포넌트 디자인 및 기능 구현

* Feat: getApi 기능 추가

* Refactor: GNB 컴포넌트 button 태그 Button 컴포넌트로 변경

* Feat: mainPage 추가 (#9)

* Fix: Button 컴포넌트 리팩토링 (#10)

* Fix: font-family 스타일 Link 불어오기 추가

* Fix: Button 컴포넌트 리팩토링

* Feat: TabList 컴포넌트 구현 (#11)

* Fix: Button 컴포넌트 리팩토링

* Fix: Button 컴포넌트 floating 타입 추가

* Fix: Button 컴포넌트 리팩토링

* Feat: TabList 컴포넌트 구현

* Feat: HeaderService 컴포넌트 디자인 및 기능 구현 (#12)

* Feat: 컴포넌트 명 변경

* Feat: HeaderService 컴포넌트 디자인 및  기능 구현

 드롭다운 버튼 클릭시 리액션 이모지 디스플레이  기능 구현

* Refactor: disabled 상태일때 버튼 아이콘 색상 변경 기능  수정

* Feat: Card 컨포넌트 디자인 및 기능 구현 (#13)

* Feat: SenderName 컨포넌트 기능 구현

* Style: checkbutton.png 이름 변경

* Feat: Card, AddCard 컨포넌트 디자인 및 기능 구현
* Feat: Card, AddCard 컨포넌트 디자인 및 기능 구현

* Feat: Button 컨포넌트에 delete HEIGHT 추가

* Refactor: Input 컨포넌트 리팩토링

* Refactor: Avatar, Badge, Date, SenderName 컨포넌트 리팩토링

* Feat: CreatePage 진행 중

* Feat: PostPage 기능 구현
*Refactor: Card 컨포넌트 리팩토링

*Feat: formatDate 추가

* Style: orange token을 beige로 변경

* fix: Button 컴포넌트 불필요한 코드 제거

* Feat: color card 컴포넌트 구현 (#14)

* Design: avatar.png 디자인 변경

* Feat: color card 컴포넌트 구현

* Fix: Avatar 컴포넌트 버그 수정

* Feat: Modal 추가 및 일부 컴포넌트 네이밍 수정 (#15)

* Feat: CardList 컴포넌트 구현 (#16)

* Design: avatar.png 디자인 변경

* Feat: color card 컴포넌트 구현

* Fix: Avatar 컴포넌트 버그 수정

* Feat: CardList 컴포넌트 구현

* Refactor: getApi 위치 변경

CardList 컴포넌트 내의 getApi를 ListPage로 위치 변경

* Feat: api 분리 및 모듈화  (#17)

* Feat: api 분리 및 모듈화

1. http 메소드에 따라 파일 분리
2. postPage, headerService 수정한 api 함수 적용

* Refactor: 모듈 api 적용 및 props 이름 수정

* Refactor: 컴포넌트 네임 수정 (#18)

* Feat: EditPage 추가 및 기타 수정 (#19)

* Feat: EditPage 추가

* Feat: EditPage 추가

* Feat: Listpage 구현 (#20)

* Feat: api 기능 추가

* Style: Card, CardList 이름 변경

* Feat: ListPage에 컴포넌트 배치 완료

ListPage내에 컴포넌트 배치

* Feat: ListPage 버튼에 기능 추가

ListPage 버튼에 post 페이지로 가는 기능 추가
Avatar, Avatars, MessageCounter, PostCard, PostCardList 스타일 선언 위치
변경

* Feat: PostCard onClick 이벤트 추가

* Refactor: onClick이벤트 Link로 수정

* Feat: ListPage 좌우버튼 구현

* Feat: ListPage CSS 적용

* Feat: PostPage 디자인 및 Modal 컨포넌트 기능 구현 (#21)

* Feat: PostPage 디자인 및 기능 구현 중

* Fix: api.jsx 코드 수정

* Fix: Modal 컨포넌트 작동 및 PostPage  디자인 수정

* Fix: PostPage 디자인 수정

* Feat: CreatePage 페이지 구현 (#23)

* Style: Icons, Button 컴포넌트 디렉토리 경로 수정

* Remove: 불필요한 이미지 소스 삭제

* Style: Button 컴포넌트, Icon.js 디렉토리 경로 수정

* Fix: React Hook Import 추가

* Style: eslint "off" 옵션 추가

* Feat: CreatePage 구현

* Remove: Option 컴포넌트 파일 삭제

* Feat: CreatePage 페이지 구현

* Feat: MessageCounter 반응형 디자인 추가 (#22)

* Refactor: 컨포넌트 파일 이동에 따른 경로 수정

* Feat: EditPage 반응형 추가 (#24)

* Feat: 이모지 라이브러리 기능 구현 (#25)

* Fix: PostPage의 background min-height로 변경

* Feat: emoji-picker 기능 추가 중

* Feat: 이모지 라이브러리 기능 구현

* Feat: message 페이지 디자인 및 기능 구현 (#26)

* Fix: 페이지 병합 및 충돌 해결

* Feat: 공유 버튼 디자인 및 기능 구현 (#27)

* Fix: 버그 수정 (#29)

* Fix: 캐시 삭제

* Feat: 공유하기 기능 구현 (#30)

* Fix: 카드, 모달 컴포넌트 font 데이터에 따른 font-family 적용 (#32)

* Fix: 오타 수정

* Feat: 반응형 스타일 구현 (#33)

* Feat: 무한 스크롤 기능 구현 (#34)

* Feat: 무한 스크롤 구현

* Feat: 편집하기 버튼 추가

* Fix: 버그 수정

* Fix: 불필요한 코드 삭제

불필요한 코드 삭제 및 eslintrc rules에 "indent": "off" 추가

* Fix: transient props(props 네임 경고) 수정

* Fix: 카카오톡 공유 url 수정

* Fix: 반응형 스타일 구현 수정 (#36)

* Fix: props 초기 타입 수정

* Fix: 반응형 스타일 구현 수정

* Fix: 이미지 적용 문제 해결 (#37)

* Fix: 카카오톡 공유 기능 수정

* Fix: 불필요한 import 제거 및 수정

* Feat: textEditor 기능 구현  (#38)

* Feat: textEditor 기능 구현

1. ckeditor5 설치
2. html paser 설치

* Design: 헬멧 타이틀 텍스트  변경

* Fix: 이미지배경 선택 오류 수정

* Fix: 카카오톡 공유 기능 이미지 변수 수정

* Fix: 카카오톡 공유하기 컴포넌트 내 title, description 값 수정

* Fix: public 폴더에 '_redirects'파일 추가

* Fix: 카카오톡 공유하기 모바일 이슈 수정

* Refator: 디자인 검수 및 반응형 스타일 구현 (#41)

* Fix: 반응형 스타일 구현 수정

* Refactor: EditPage 디자인 검수 및 반응형 스타일 구현

* Fix: CreatePage 옵션 이미지 반응형 수정

* Refator: 디자인 검수 및 반응형 스타일 구현

* Fix: 카카오톡 공유 웹 경로 추가

* Fix: 카카오톡 공유 기능 불필요한  콘솔부분 제거

* Feat: 파비콘 추가

* Refactor: 반응형 수정 (#43)

* Feat: Skeleton 구현 (#45)

* Feat: ListPage Skeleton 기능 구현

* Feat: PostPage Skeleton 기능 구현

* Design: Skeleton 디자인 변경

* Fix: 무한 스크롤 버그 수정

* Refactor: InputText 컴포넌트 교체 (#46)

* Refactor: ListPage 반응형 수정 (#47)

* Doc: 프로젝트 READEME.md 초안 작성

* Update README.md

* Update README.md

* Fix: 반응형 수정

* Fix: 컴포넌트 내 스타일 컴포넌트 위치 변경

* Update README.md

* Design: 크리스마스 디자인 적용 (#48)

* Fix: 중복 스타일 컴포넌트 제거

* Fix: 중복 스타일 컴포넌트 제거 (#51)

---------

+++ src/pages/EditPage.jsx
@@ -48,8 +48,13 @@ function EditPage() {
   const bgImg = bgData.backgroundImageURL;
 
   const handleDeleteRecipients = async () => {
-    await deleteRecipientDAsync(id);
-    navigate('/list');
+    // eslint-disable-next-line no-alert
+    const userConfirmed = window.confirm('정말로 삭제하시겠습니까?');
+
+    if (userConfirmed) {
+      await deleteRecipientDAsync(id);
+      navigate('/list');
+    }
   };
 
   return (
@@ -91,7 +96,7 @@ const ContentWrapper = styled.div`
   margin: 0 auto;
   padding: 6.3rem 2.4rem;
 
-  @media (max-width: ${layout.breakpoint.mobile}) {
+  @media (max-width: ${layout.breakpoint.tablet}) {
     padding: 2.4rem 2.4rem 10.4rem;
   }
 `;
@@ -102,14 +107,14 @@ const StyledButton = styled.div`
   width: 100%;
   margin-bottom: 1.1rem;
 
-  @media (max-width: ${layout.breakpoint.mobile}) {
+  @media (max-width: ${layout.breakpoint.tablet}) {
     position: fixed;
     left: 0;
     right: 0;
     bottom: 2.4rem;
     padding: 0 2.4rem;
     margin-bottom: 0;
-
+  
     button {
       width: 100%;
       height: 5.6rem;
@@ -126,7 +131,7 @@ const Container = styled.div`
       return `url(${props.$bgImg})`;
     }
     if (props.$bgColor) {
-      return color[props.$bgColor][200];
+      return color.xMas[props.$bgColor][200];
     }
     return color.white;
   }};
